align vesselDraft & milesRemainingToDestination types

+++ ui-support-persistence/src/main/java/org/dcsa/uisupport/persistence/entity/TransportCallWithTimestamps.java
@@ -31,8 +31,8 @@ public class TransportCallWithTimestamps extends TransportCall {
   private OffsetDateTime atdBerthDateTime;
 
   @Column(name = "vessel_draft")
-  private Integer vesselDraft;
+  private Float vesselDraft;
 
   @Column(name = "miles_remaining_to_destination")
-  private Integer milesRemainingToDestination;
+  private Float milesRemainingToDestination;
 }
